+++ CHANGELOG.md
@@ -21,6 +21,18 @@
 - Fixes the typespec for command dispatch ([#325](https://github.com/commanded/commanded/pull/325)).
 - Process manager stops if `interested?/1` returns an empty list ([#335](https://github.com/commanded/commanded/pull/335)).
 
+## v1.0.1
+
+### Enhancements
+
+- Global registry using Erlang's `:global` module ([#344](https://github.com/commanded/commanded/pull/344)).
+- Command dispatch return ([#331](https://github.com/commanded/commanded/pull/331)).
+
+### Bug fixes
+
+- Fix distributed subscription registration bug ([#345](https://github.com/commanded/commanded/pull/345)).
+- Retry event handler and process manager suscriptions on error ([#348](https://github.com/commanded/commanded/pull/348)).
+
 ## v1.0.0
 
 ### Breaking changes